refactor: move all pages to pages folder
