+++ frontend/engetec-app/src/app/app-routing.module.ts
@@ -3,15 +3,15 @@ import { RouterModule, Routes } from '@angular/router';
 
 const routes: Routes = [
   { path: '', pathMatch: 'full', redirectTo: '/login' },
-  { path: 'login', loadChildren: () => import('./login/login.module').then(m => m.LoginModule) },
-  { path: 'register', loadChildren: () => import('./register/register.module').then(m => m.RegisterModule) },
-  { path: 'forgotpass', loadChildren: () => import('./forgotpass/forgotpass.module').then(m => m.ForgotpassModule) },
-  { path: 'forgotpassmessage', loadChildren: () => import('./forgotpass-message/forgotpass-message.module').then(m => m.ForgotpassMessageModule) },
-  { path: 'resetpass', loadChildren: () => import('./resetpass/resetpass.module').then(m => m.ResetpassModule) },
-  { path: 'dashboard', loadChildren: () => import('./dashboard/dashboard.module').then(m => m.DashboardModule) },
-  { path: 'novo-artigo', loadChildren: () => import('./artigo/artigo.module').then(m => m.ArtigoModule) },
-  { path: 'atribuir', loadChildren: () => import('./atribuicao/atribuicao.module').then(m => m.AtribuicaoModule) },
-  { path: 'avaliar', loadChildren: () => import('./avaliacao/avaliacao.module').then(m => m.AvaliacaoModule) }
+  { path: 'login', loadChildren: () => import('./pages/login/login.module').then(m => m.LoginModule) },
+  { path: 'register', loadChildren: () => import('./pages/register/register.module').then(m => m.RegisterModule) },
+  { path: 'forgotpass', loadChildren: () => import('./pages/forgotpass/forgotpass.module').then(m => m.ForgotpassModule) },
+  { path: 'forgotpassmessage', loadChildren: () => import('./pages/forgotpass-message/forgotpass-message.module').then(m => m.ForgotpassMessageModule) },
+  { path: 'resetpass', loadChildren: () => import('./pages/resetpass/resetpass.module').then(m => m.ResetpassModule) },
+  { path: 'dashboard', loadChildren: () => import('./pages/dashboard/dashboard.module').then(m => m.DashboardModule) },
+  { path: 'novo-artigo', loadChildren: () => import('./pages/artigo/artigo.module').then(m => m.ArtigoModule) },
+  { path: 'atribuir', loadChildren: () => import('./pages/atribuicao/atribuicao.module').then(m => m.AtribuicaoModule) },
+  { path: 'avaliar', loadChildren: () => import('./pages/avaliacao/avaliacao.module').then(m => m.AvaliacaoModule) }
 ];
 
 @NgModule({
