+++ frontend/engetec-app/src/app/pages/artigo/artigo-routing.module.ts
