+++ frontend/engetec-app/src/app/pages/artigo/artigo.component.ts
