+++ frontend/engetec-app/src/app/pages/artigo/artigo.module.ts
