+++ frontend/engetec-app/src/app/pages/dashboard/dashboard.component.ts
