+++ frontend/engetec-app/src/app/pages/forgotpass-message/forgotpass-message.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AngularMaterialModule } from '../angular-material/angular-material.module';
+import { AngularMaterialModule } from '../../angular-material/angular-material.module';
 
 import { ForgotpassMessageRoutingModule } from './forgotpass-message-routing.module';
 import { ForgotpassMessageComponent } from './forgotpass-message.component';
