+++ frontend/engetec-app/src/app/pages/forgotpass/forgotpass.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AngularMaterialModule } from '../angular-material/angular-material.module';
+import { AngularMaterialModule } from '../../angular-material/angular-material.module';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { ForgotpassRoutingModule } from './forgotpass-routing.module';
