+++ frontend/engetec-app/src/app/pages/register/register.component.scss
