+++ frontend/engetec-app/src/app/pages/register/register.component.ts
@@ -1,9 +1,9 @@
 import { Component } from '@angular/core';
 import { FormControl, FormGroup, Validators } from '@angular/forms';
-import { AuthService } from '../_services/auth/auth.service';
+import { AuthService } from '../../_services/auth/auth.service';
 import { Router } from '@angular/router';
 import { first } from 'rxjs/operators';
-import { passwordMatchValidator } from '../_shared/password/password-match.directive'
+import { passwordMatchValidator } from '../../_shared/password/password-match.directive'
 
 @Component({
   selector: 'app-register',
